requirements.txt
alpaca-py>=0.36.0
yfinance>=0.2.40
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.32.0
apscheduler>=3.10.0
anthropic>=0.45.0
